% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFplot.R
\name{CDFplot}
\alias{CDFplot}
\title{Cumulative Distribution Function Plotting}
\usage{
CDFplot(SEresultlist, efficiency_5ss_3ss_nona_inf_reduct, zoom.x = c(5, 6))
}
\arguments{
\item{SEresultlist}{a large list created by SEcalculation}

\item{efficiency_5ss_3ss_nona_inf_reduct}{a dataframe created by SEfilter}

\item{zoom.x}{a numeric vector with 2 elements}
}
\value{
a large list with K-S test result
}
\description{
Plot the Cumulative Distribution Function of splicing efficiency.
}
\examples{
\donttest{
CDFplot(SEresultlist,efficiency_5ss_3ss_nona_inf_reduct,zoom.x= c(5,6))
}

}
