% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_5ss_3ss_nona_inf_reduct.R
\docType{data}
\name{efficiency_5ss_3ss_nona_inf_reduct}
\alias{efficiency_5ss_3ss_nona_inf_reduct}
\title{Some random data}
\format{
A dataframe
}
\source{
Generated from SEAA data
}
\description{
A few numbers from \code{efficiency_5ss_3ss_nona_inf_reduct}.
}
\examples{
data(efficiency_5ss_3ss_nona_inf_reduct)
str(efficiency_5ss_3ss_nona_inf_reduct)
}
