% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEfilter.R
\name{SEfilter}
\alias{SEfilter}
\title{Splicing efficiency Filtering}
\usage{
SEfilter(SEresultlist, min_counts = 5)
}
\arguments{
\item{SEresultlist}{a large list created by SEcalculation}

\item{min_counts}{a numeric value}
}
\value{
a dataframe
}
\description{
Unconventional splicing efficiency will be removed and reads counts less than min_counts will be filtered.
}
\examples{
\donttest{
efficiency_5ss_3ss_nona_inf_reduct<-SEfilter(SEresultlist,min_counts = 5)
}
}
