% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEresultlist.R
\docType{data}
\name{SEresultlist}
\alias{SEresultlist}
\title{Some random data}
\format{
A large list
}
\source{
Generated from SEAA data
}
\description{
A few numbers from \code{SEresultlist}.
}
\examples{
data(SEresultlist)
str(SEresultlist)
}
