% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEsiteanno.R
\name{SEsiteanno}
\alias{SEsiteanno}
\title{Annotation of Splicing Sites Acquiring Filtered Splicing Efficiency}
\usage{
SEsiteanno(efficiency_5ss_3ss_nona_inf_reduct)
}
\arguments{
\item{efficiency_5ss_3ss_nona_inf_reduct}{a data frame created by SEfilter}
}
\value{
a data frame
}
\description{
This function helps to annotate the splicing sites with clusterProfiler. The gene symbol, REFSEQ, ENSEMBL are shown.
}
\examples{
\donttest{
SEannotaionresult<-SEsiteanno(efficiency_5ss_3ss_nona_inf_reduct)
}
}
