% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEtypeplot.R
\name{SEtypeplot}
\alias{SEtypeplot}
\title{Splicing Status Type Plotting}
\usage{
SEtypeplot(SEresultlist, arrangement)
}
\arguments{
\item{SEresultlist}{a large list created by SEcalculation}

\item{arrangement}{a characteristic value}
}
\value{
a large list
}
\description{
Splicing Status Type Plotting
}
\examples{
\donttest{
SEtyperesult<-SEtypeplot(SEresultlist,"horizontal")
}

}
